Extract replay tick mark normalization
