+++ js/player/src/replay-data.ts
@@ -5,7 +5,6 @@ import type {
   RawDemolishInfo,
   RawGoalEvent,
   RawPlayerStatEvent,
-  RawReplayTickMark,
   PlayerSample,
   RawBallFrame,
   RawPlayerFrame,
@@ -25,6 +24,7 @@ import {
   STANDARD_SOCCAR_BOOST_PAD_COUNT,
 } from "./replay-boost-pads";
 import { buildPlayerLookup, normalizeReplayTime, playerIdToString } from "./replay-data-helpers";
+import { buildReplayTickMarks, replayTickMarkTimelineEvent } from "./replay-tick-marks";
 
 export interface NormalizeReplayDataOptions {
   onProgress?: (progress: number, details: NormalizeReplayProgress) => void;
@@ -697,66 +697,6 @@ async function buildBallFramesAsync(
 
 function createTimelineEventId(prefix: string, frame: number, suffix: string): string {
   return `${prefix}:${frame}:${suffix}`;
-}
-
-function tickMarkFrame(tickMark: RawReplayTickMark): number | null {
-  return Number.isInteger(tickMark.frame) && tickMark.frame >= 0 ? tickMark.frame : null;
-}
-
-function tickMarkRawTime(tickMark: RawReplayTickMark, raw: RawReplayFramesData): number | null {
-  if (typeof tickMark.time === "number" && Number.isFinite(tickMark.time)) {
-    return tickMark.time;
-  }
-
-  const frame = tickMarkFrame(tickMark);
-  if (frame === null) {
-    return null;
-  }
-
-  const frameTime = raw.frame_data.metadata_frames[frame]?.time;
-  return typeof frameTime === "number" && Number.isFinite(frameTime) ? frameTime : null;
-}
-
-function replayTickMarkId(tickMark: RawReplayTickMark, index: number): string {
-  const frame = tickMarkFrame(tickMark);
-  return `bookmark:${frame ?? "unknown"}:${tickMark.description || "tick-mark"}:${index}`;
-}
-
-function buildReplayTickMarks(
-  raw: RawReplayFramesData,
-  startTime: number,
-  progressTracker?: NormalizeReplayProgressTracker,
-): ReplayTickMark[] {
-  return (raw.replay_tick_marks ?? []).flatMap((tickMark, index) => {
-    progressTracker?.advance();
-    const rawTime = tickMarkRawTime(tickMark, raw);
-    if (rawTime === null) {
-      return [];
-    }
-
-    return [
-      {
-        id: replayTickMarkId(tickMark, index),
-        description: tickMark.description,
-        frame: tickMarkFrame(tickMark),
-        time: normalizeReplayTime(rawTime, startTime),
-      },
-    ];
-  });
-}
-
-function replayTickMarkTimelineEvent(tickMark: ReplayTickMark): ReplayTimelineEvent {
-  const label = tickMark.description.trim() || "Replay bookmark";
-  return {
-    id: tickMark.id,
-    time: tickMark.time,
-    seekTime: tickMark.time,
-    frame: tickMark.frame ?? undefined,
-    kind: "bookmark",
-    label,
-    shortLabel: "BM",
-    iconName: "bookmark",
-  };
 }
 
 function sortTimelineEvents(events: ReplayTimelineEvent[]): ReplayTimelineEvent[] {
